OnlineShop/requirements.txt
django==5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
drf-spectacular
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
python-dotenv==1.0.1
